requirements.txt
# Core ML/DL dependencies
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
tokenizers

# Data science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Chemistry
rdkit

# Configuration and experiment tracking
omegaconf>=2.3.0
mlflow>=2.0.0
types-PyYAML>=6.0.0

# Utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.60.0
rich>=13.0.0

# Development tools (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Code quality (optional)
mypy>=1.0.0
black>=22.0.0
isort>=5.10.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Additional utilities
hydra-core>=1.3.0
tensorboard>=2.10.0
